Remove New Length Counter For Temp Array
Forgot to remove this code - The new length counter for the temp array was never used. The initial idea was to make the length of the temp array the length of the new password. However, since it is a temp array we can just keep the old length because the new length is guaranteed to not be greater than the old length.