added character repetition penalty
By adding character repetition for 93 ASCII printable characters (excludes space and comma), we penalize for character repetition in the evaluation of the password.
For all 93 characters we include a string that repeats the character 4 times and a string that repeats the character 7 times. The algorithm will detect fuzz with one edit distance, so the 4-character string will pick up fuzz if a character is repeated 3, 4, or 5 times. The 7-character string will pick up fuzz if a character is repeated 6, 7, or 8 times. Any greater than that would just get picked up again as another penalty. For example: 'AAAA' in the list would pick up 'AAAB' in a password evaluation.